Split typechecker resolver root tests
