+++ src/typechecker/tests/resolver_impl_values.rs
@@ -1,0 +1,276 @@
+use super::*;
+
+#[test]
+fn check_program_with_symbols_requires_resolver_impl_methods() {
+    let program = parse_program(
+        r#"
+Json: behavior {
+    stringify: (Self) str
+}
+
+Point: { x: i32 }
+
+Point.implements(Json) {
+    stringify = (value: Point) str { "point" }
+}
+"#,
+    );
+    let symbols = SymbolTable::default();
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("missing impl method resolver symbols should fail");
+
+    assert!(
+        err.iter().any(|d| d
+            .message
+            .contains("resolver symbol table missing value symbol 'Point.stringify'")),
+        "expected missing impl method symbol diagnostic, got {err:?}"
+    );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_impl_method_signature() {
+    let program = parse_program(
+        r#"
+Json: behavior {
+    stringify: (Self) str
+}
+
+Point: { x: i32 }
+
+Point.implements(Json) {
+    stringify = (value: Point) str { "point" }
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_return_type_name_for_test(
+        Namespace::Value,
+        "Point.stringify",
+        Some("i32".to_string()),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver impl method signature mismatch should fail");
+
+    assert!(
+        err.iter().any(|d| d.message.contains(
+            "resolver value symbol 'Point.stringify' has return type 'i32', expected 'str'"
+        )),
+        "expected resolver impl method signature diagnostic, got {err:?}"
+    );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_impl_function_type_signature() {
+    let program = parse_program(
+        r#"
+Mapper: behavior {
+    map: (Self, (i32) i32) (i32) i32
+}
+
+Point: { x: i32 }
+
+Point.implements(Mapper) {
+    map = (value: Point, callback: (i32) i32) (i32) i32 {
+        callback
+    }
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_return_type_name_for_test(Namespace::Value, "Point.map", Some("i32".to_string()));
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver impl method function type mismatch should fail");
+
+    assert!(
+        err.iter().any(|d| d.message.contains(
+            "resolver value symbol 'Point.map' has return type 'i32', expected '(i32) i32'"
+        )),
+        "expected resolver impl method function type diagnostic, got {err:?}"
+    );
+}
+
+#[test]
+fn check_program_with_symbols_requires_resolver_impl_method_body_locals() {
+    let program = parse_program(
+        r#"
+Json: behavior {
+    stringify: (Self) str
+}
+
+Point: { x: i32 }
+
+Point.implements(Json) {
+    stringify = (value: Point) str {
+        label = "point"
+        label
+    }
+}
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.remove_for_test(Namespace::Local, "label");
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("missing resolver impl method body local should fail");
+
+    assert!(
+        err.iter().any(|d| d
+            .message
+            .contains("resolver symbol table missing local symbol 'label'")),
+        "expected missing resolver impl method body local diagnostic, got {err:?}"
+    );
+}
+
+#[test]
+fn check_program_with_symbols_requires_resolver_enum_variants() {
+    let program = parse_program(
+        r#"
+Option: Some(i32), None
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.remove_for_test(Namespace::Variant, "Some");
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("missing resolver enum variant symbols should fail");
+
+    assert!(
+        err.iter().any(|d| d
+            .message
+            .contains("resolver symbol table missing variant symbol 'Some'")),
+        "expected missing enum variant symbol diagnostic, got {err:?}"
+    );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_function_arity() {
+    let program = parse_program(
+        r#"
+add = (a: i32, b: i32) i32 { a + b }
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_parameter_count_for_test(Namespace::Value, "add", Some(1));
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver function arity mismatch should fail");
+
+    assert!(
+        err.iter().any(|d| d
+            .message
+            .contains("resolver value symbol 'add' has parameter count 1, expected 2")),
+        "expected resolver function arity diagnostic, got {err:?}"
+    );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_function_parameter_types() {
+    let program = parse_program(
+        r#"
+add = (a: i32, b: f64) f64 { b }
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_parameter_type_names_for_test(
+        Namespace::Value,
+        "add",
+        Some(vec!["i32".to_string(), "i32".to_string()]),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver function parameter type mismatch should fail");
+
+    assert!(
+        err.iter().any(|d| d.message.contains(
+            "resolver value symbol 'add' has parameter types '(i32, i32)', expected '(i32, f64)'"
+        )),
+        "expected resolver function parameter type diagnostic, got {err:?}"
+    );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_function_type_parameter_metadata() {
+    let program = parse_program(
+        r#"
+apply = (callback: (i32) i32, value: i32) i32 { value }
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_parameter_type_names_for_test(
+        Namespace::Value,
+        "apply",
+        Some(vec!["i32".to_string(), "i32".to_string()]),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver function type parameter metadata mismatch should fail");
+
+    assert!(
+            err.iter().any(|d| d.message.contains(
+                "resolver value symbol 'apply' has parameter types '(i32, i32)', expected '((i32) i32, i32)'"
+            )),
+            "expected resolver function type parameter metadata diagnostic, got {err:?}"
+        );
+}
+
+#[test]
+fn check_program_with_symbols_validates_resolver_function_parameter_names() {
+    let program = parse_program(
+        r#"
+add = (a: i32, b: f64) f64 { b }
+"#,
+    );
+    let mut symbols = crate::resolver::Resolver::new()
+        .resolve_program(&program)
+        .expect("resolver succeeds");
+    symbols.set_parameter_names_for_test(
+        Namespace::Value,
+        "add",
+        Some(vec!["a".to_string(), "other".to_string()]),
+    );
+    let mut tc = TypeChecker::new();
+
+    let err = tc
+        .check_program_with_symbols(&program, &symbols)
+        .expect_err("resolver function parameter name mismatch should fail");
+
+    assert!(
+        err.iter().any(|d| d.message.contains(
+            "resolver value symbol 'add' has parameter names '(a, other)', expected '(a, b)'"
+        )),
+        "expected resolver function parameter name diagnostic, got {err:?}"
+    );
+}
